home visivel

+++ app/src/main/java/br/com/rodriguesalex/moviedb/models/MDbMovie.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 class MDbMovie {
     @SerializedName("poster_path") val poster: String? = null
     @SerializedName("title") val title: String? = null
-    @SerializedName("vote_average") val voteAverage: Int? = null
+    @SerializedName("vote_average") val voteAverage: Double? = null
     @SerializedName("release_date") val releaseDate: String? = null
 }
